server/requirements.txt
Django>=3.2,<4.0
psycopg2-binary>=2.9.1
redis>=4.0.0
django_redis==5.2.0
Pillow
python-decouple
djangorestframework